% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_addresses_sf}
\alias{nyc_addresses_sf}
\title{NYCHA residential addresses (spatial)}
\format{
\subsection{\code{nyc_addresses_sf}}{

A simple feature collection with one row per geocoded address and
all of the columns in \link{nyc_addresses_df} except \code{latitude} and
\code{longitude}, which are replaced by a \code{geometry} column.
\describe{
\item{geometry}{Point geometry in EPSG:2263.}
}
}
}
\source{
NYC OpenData: \url{https://data.cityofnewyork.us/Housing-Development/NYCHA-Residential-Addresses/k6ya-3sk4}
}
\usage{
nyc_addresses_sf
}
\description{
Point geometries for the NYCHA residential addresses in
\link{nyc_addresses_df}. Rows missing latitude or longitude are dropped.
Coordinates reprojected to EPSG:2263, NAD83 / New York Long Island
(ftUS).
}
\author{
Kieran Healy
}
\keyword{datasets}
