% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_community_facilities_sf}
\alias{nyc_community_facilities_sf}
\title{Directory of NYCHA community facilities (spatial)}
\format{
\subsection{\code{nyc_community_facilities_sf}}{

A simple feature collection with one row per geocoded facility and
all of the columns in \link{nyc_community_facilities_df} except \code{latitude}
and \code{longitude}, which are replaced by a \code{geometry} column.
\describe{
\item{geometry}{Point geometry in EPSG:2263.}
}
}
}
\source{
NYC OpenData: \url{https://data.cityofnewyork.us/Housing-Development/Directory-Of-NYCHA-Community-Facilities/r6n9-zw2y}
}
\usage{
nyc_community_facilities_sf
}
\description{
Point geometries for the NYCHA community facilities in
\link{nyc_community_facilities_df}. Rows missing latitude or longitude
are dropped. Coordinates reprojected to EPSG:2263, NAD83 / New York
Long Island (ftUS).
}
\author{
Kieran Healy
}
\keyword{datasets}
